Disable self super admin account deletion

+++ src/Controller/UserSelfDeleteAccount.php
@@ -3,9 +3,12 @@
 namespace App\Controller;
 
 use App\Entity\User;
+use App\Enum\UserRole;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\HttpException;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class UserSelfDeleteAccount extends AbstractController {
@@ -14,6 +17,10 @@ class UserSelfDeleteAccount extends AbstractController {
     $user = $this->getUser();
     if (!$user instanceof User) {
       throw new NotFoundHttpException();
+    }
+
+    if ($user->getRole() === UserRole::super_admin) {
+      throw new HttpException(Response::HTTP_BAD_REQUEST, "You can't delete this account.");
     }
 
     $em->remove($user);
